@@ -71,14 +71,13 @@
     "adldap2/adldap2-laravel": "6.*",
     "bacon/bacon-qr-code": "1.*",
     "bunq/sdk_php": "dev-master",
-    "danhunsaker/laravel-flysystem-others": "1.*",
     "davejamesmiller/laravel-breadcrumbs": "5.*",
     "doctrine/dbal": "2.*",
     "fideloper/proxy": "4.*",
     "jc5/google2fa-laravel": "2.0.3",
-    "laravel/framework": "5.8.*",
-    "laravel/passport": "7.*",
-    "laravelcollective/html": "5.8.*",
+    "laravel/framework": "^6.0",
+    "laravel/passport": "8.*",
+    "laravelcollective/html": "6.*",
     "league/commonmark": "1.*",
     "league/csv": "9.*",
     "league/flysystem-replicate-adapter": "1.*",
@@ -86,16 +85,16 @@
     "league/fractal": "0.*",
     "litipk/flysystem-fallback-adapter": "0.*",
     "mschindler83/fints-hbci-php": "1.*",
-    "pragmarx/google2fa": "6.1.3",
+    "pragmarx/google2fa": "^7.0",
     "pragmarx/recovery": "^0.1.0",
     "predis/predis": "^1.1",
-    "rcrowe/twigbridge": "0.9.*"
+    "rcrowe/twigbridge": "^0.11.2"
   },
   "require-dev": {
     "barryvdh/laravel-ide-helper": "2.*",
     "filp/whoops": "2.*",
     "fzaninotto/faker": "1.*",
-    "johnkary/phpunit-speedtrap": "^3.0",
+    "johnkary/phpunit-speedtrap": "^3.1",
     "mockery/mockery": "1.*",
     "phpunit/phpunit": "8.*",
     "roave/security-advisories": "dev-master"
